selected motifs closest to centroids

+++ subsample-filters.py
@@ -8,6 +8,7 @@ import numpy as np
 import os
 import pandas as pd
 import pickle
+from scipy.spatial.distance import cosine
 from sklearn.cluster import KMeans
 import subprocess as sp
 
@@ -36,57 +37,54 @@ def main(**params):
 
     # Initialize
     base_dir = os.path.dirname(os.path.realpath(__file__))
-    data_dir = os.path.join(base_dir, "data")
     prefix = os.path.splitext(os.path.basename(params["binding_modes_pickle"]))[0]
 
     # Load
-    probs = {}
     matrix_id2filter = {}
-    ixs = {"A":0, "C":1, "G":2, "T":3}
     with open(params["binding_modes_pickle"], "rb") as handle:
         filters = pickle.load(handle)
+    for bm in filters:
+        name = "%s;fwd;%s" % (filters[bm][0][0], "::".join(filters[bm][0][1]))
+        matrix_id2filter.setdefault(name, filters[bm][0][2])
+        name = "%s;rev;%s" % (filters[bm][0][0], "::".join(filters[bm][0][1]))
+        matrix_id2filter.setdefault(name, filters[bm][0][3])
 
     meme_file = os.path.join(base_dir, "%s.meme" % prefix)
+    if not os.path.exists(meme_file):
 
-    lines = ""
-    lines += "MEME version 4\n\n"
-    lines += "ALPHABET= ACGT\n\n"
-    lines += "strands: + -\n\n"
-    lines += "Background letter frequencies (from uniform background):\n"
-    lines += "A 0.25000 C 0.25000 G 0.25000 T 0.25000\n\n"
+        lines = ""
+        lines += "MEME version 4\n\n"
+        lines += "ALPHABET= ACGT\n\n"
+        lines += "strands: + -\n\n"
+        lines += "Background letter frequencies (from uniform background):\n"
+        lines += "A 0.25000 C 0.25000 G 0.25000 T 0.25000\n\n"
 
-    for bm in sorted(filters):
-        m = "\n".join(["\t".join(list(map(str, i+.25))) for i in filters[bm][0][2]])
-        handle = io.StringIO(m)
-        m = motifs.read(handle, "pfm-four-columns")
-        name = "%s;fwd;%s" % (filters[bm][0][0], "::".join(filters[bm][0][1]))
-        consensus_seq = m.consensus
-        w = len(consensus_seq)
-        lines += "MOTIF %s %s\n" % (name, consensus_seq)
-        lines += "letter-probability matrix: alength= 4 w= %s nsites= 20 E= 0\n" % w
-        for row in np.transpose(np.array(list(m.pwm.values()))):
-            lines += " ".join([str(round(r, 8)).rjust(11) for r in row]) + "\n"
-        lines += "\n"
-        matrix_id2filter.setdefault(name, filters[bm][0][2])
-        prob = sum([m.pwm[ixs[consensus_seq[i]]][i] for i in range(len(consensus_seq))])
-        probs.setdefault(name, prob)
-        m = "\n".join(["\t".join(list(map(str, i+.25))) for i in filters[bm][0][3]])
-        handle = io.StringIO(m)
-        m = motifs.read(handle, "pfm-four-columns")
-        name = "%s;rev;%s" % (filters[bm][0][0], "::".join(filters[bm][0][1]))
-        consensus_seq = m.consensus
-        w = len(consensus_seq)
-        lines += "MOTIF %s %s\n" % (name, consensus_seq)
-        lines += "letter-probability matrix: alength= 4 w= %s nsites= 20 E= 0\n" % w
-        for row in np.transpose(np.array(list(m.pwm.values()))):
-            lines += " ".join([str(round(r, 8)).rjust(11) for r in row]) + "\n"
-        lines += "\n"
-        matrix_id2filter.setdefault(name, filters[bm][0][3])
-        prob = sum([m.pwm[ixs[consensus_seq[i]]][i] for i in range(len(consensus_seq))])
-        probs.setdefault(name, prob)
+        for bm in sorted(filters):
+            m = "\n".join(["\t".join(list(map(str, i+.25))) for i in filters[bm][0][2]])
+            handle = io.StringIO(m)
+            m = motifs.read(handle, "pfm-four-columns")
+            name = "%s;fwd;%s" % (filters[bm][0][0], "::".join(filters[bm][0][1]))
+            consensus_seq = m.consensus
+            w = len(consensus_seq)
+            lines += "MOTIF %s %s\n" % (name, consensus_seq)
+            lines += "letter-probability matrix: alength= 4 w= %s nsites= 20 E= 0\n" % w
+            for row in np.transpose(np.array(list(m.pwm.values()))):
+                lines += " ".join([str(round(r, 8)).rjust(11) for r in row]) + "\n"
+            lines += "\n"
+            m = "\n".join(["\t".join(list(map(str, i+.25))) for i in filters[bm][0][3]])
+            handle = io.StringIO(m)
+            m = motifs.read(handle, "pfm-four-columns")
+            name = "%s;rev;%s" % (filters[bm][0][0], "::".join(filters[bm][0][1]))
+            consensus_seq = m.consensus
+            w = len(consensus_seq)
+            lines += "MOTIF %s %s\n" % (name, consensus_seq)
+            lines += "letter-probability matrix: alength= 4 w= %s nsites= 20 E= 0\n" % w
+            for row in np.transpose(np.array(list(m.pwm.values()))):
+                lines += " ".join([str(round(r, 8)).rjust(11) for r in row]) + "\n"
+            lines += "\n"
 
-    with open(meme_file, "w") as handle:
-        handle.write(lines)
+        with open(meme_file, "w") as handle:
+            handle.write(lines)
 
     tomtom_dir = os.path.join(base_dir, prefix)
     if not os.path.isdir(tomtom_dir):
@@ -108,17 +106,22 @@ def main(**params):
         ys.append(matrix_id)
     Xs = np.array(Xs)
 
-    # Get clusters
+    # K-means
     clusters = {}
-    kmeans = KMeans(n_clusters=params["n_filters"], random_state=0).fit(Xs)
-    for z in zip(kmeans.labels_, ys):
-        clusters.setdefault(z[0], [])
-        clusters[z[0]].append(z[1])
+    distances = {}
+    km = KMeans(n_clusters=params["n_filters"], random_state=0).fit(Xs)
+    zipped_list = list(zip(km.labels_, Xs, ys))
+    for z in range(len(zipped_list)):
+        cluster, X, y = zipped_list[z]
+        clusters.setdefault(cluster, [])
+        clusters[cluster].append(y)
+        distance = cosine(km.cluster_centers_[cluster], X)
+        distances.setdefault(y, distance)
 
     # Subsampled filters
     sub_filters = {}
     for c in sorted(clusters):
-        clusters[c].sort(key=lambda x: probs[x], reverse=True)
+        clusters[c].sort(key=lambda x: distances[x])
         matrix_id = clusters[c][0]
         sub_filters.setdefault(matrix_id, matrix_id2filter[matrix_id])
